+++ packages/codemirror-block-editor/src/index.ts
@@ -1,4 +1,4 @@
-export { blockLevelExtension } from "./extension/extension";
+export { blockEditor } from "./extension/extension";
 export { blockLevelListenerFacet } from "./extension/block-level-listener";
 export * from "./extension/set-block-level-effect";
 export * from "./version.gen";